Faiss/requirements.txt
faiss-cpu
streamlit
sentence_transformers